[fix] Move BaseTechObject saving in prg lua methods to other class
